Fixed issue where column had no decimal precision. Updated playground rows for fastest_laps

+++ F1Info/MockDataGenerator/src/main/java/mock_data_generator/tables/02_f1_info_tables.sql
@@ -132,7 +132,7 @@ create table fastest_laps(
   id int not null auto_increment,
   lap_achieved int,
   lap_rank int,
-  time_in_seconds int,
+  time_in_seconds float,
   display_time varchar(255),
   speed_unit varchar(3),
   speed varchar(255),
